9장 마무리

+++ jwp-basic-gradle/src/main/java/controller/qna/ApiDeleteAnswerController.java
@@ -2,7 +2,10 @@ package controller.qna;
 
 import controller.Controller;
 import dao.AnswerDao;
+import model.Answer;
 import model.Result;
+import model.User;
+import util.SessionUserUtils;
 import view.JsonView;
 import view.ModelAndView;
 
@@ -13,11 +16,22 @@ public class ApiDeleteAnswerController implements Controller {
 
     @Override
     public ModelAndView execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        Long answerId = Long.parseLong(req.getParameter("answerId"));
+        if(!SessionUserUtils.isLoggedIn(req.getSession())){
+            return new ModelAndView(new JsonView()).addObject("message", "로그인이 필요한 서비스입니다.");
+        }
+
+        long answerId = Long.parseLong(req.getParameter("answerId"));
 
         AnswerDao answerDao = new AnswerDao();
-        answerDao.delete(answerId);
+        Answer answer = answerDao.findById(answerId);
 
-        return new ModelAndView(new JsonView()).addObject("result", Result.ok()) ;
+        User user = SessionUserUtils.getUserFromSession(req.getSession());
+
+        if(user.isSameUser(answer.getWriter())){
+            answerDao.delete(answerId);
+            return new ModelAndView(new JsonView()).addObject("result", Result.ok()) ;
+        }
+
+        return new ModelAndView(new JsonView()).addObject("message", "해당 댓글 작성자만 삭제할 수 있습니다.");
     }
 }
